+++ lib/authentication/user_data.dart
@@ -1,7 +1,0 @@
-class UserData {
-  final String uid;
-  String? pfpBase64;
-
-  // constructor
-  UserData({required this.uid, required this.pfpBase64});
-}