+++ lib/user/user_data.dart
@@ -1,0 +1,14 @@
+class UserData {
+  final String uid;
+  String? pfpBase64;
+  int level;
+  int expPoints;
+
+  // constructor
+  UserData({
+    required this.uid,
+    this.pfpBase64,
+    this.level = 1,
+    this.expPoints = 0,
+  });
+}